feat(conversation): add complete conversation management system
- Implement conversation models, repositories, and PostgreSQL
  integration
- Add conversation API endpoints (get conversations, get messages,
  create conversation)
- Integrate conversation service with dependency injection
- Fix Rust edition from 2024 to 2021 for compatibility
- Add participant and message management features
- Fix middleware authorization typo and improve code formatting

+++ src/modules/conversation/schema.rs
@@ -22,7 +22,7 @@ pub struct ConversationEntity {
 pub struct PartacipantEntity {
     pub conversation_id: Uuid,
     pub user_id: Uuid,
-    pub unread_count: usize,
+    pub unread_count: i32,
     pub joined_at: chrono::DateTime<chrono::Utc>,
     pub departed_at: Option<chrono::DateTime<chrono::Utc>>,
 }
